+++ react-demo-app/src/hooks/useSdk.ts
@@ -1,8 +1,8 @@
 import {
-  GenesysCloudWebrtcSdk,
+  GenesysCloudWebrtcSdk, IExtendedMediaSession,
   ISdkConfig,
-  ISdkConversationUpdateEvent,
-  SessionTypes
+  ISdkConversationUpdateEvent, ISdkGumRequest, SdkMediaStateWithType,
+  SessionTypes, ISessionIdAndConversationId, MemberStatusMessage, VideoMediaSession
 } from 'genesys-cloud-webrtc-sdk';
 import { v4 } from 'uuid';
 import { useDispatch } from 'react-redux';
@@ -15,7 +15,15 @@ import {
 import { setSdk } from '../features/sdkSlice';
 import { updateGumRequests, updateMediaState } from '../features/devicesSlice';
 import { useSelector } from 'react-redux';
-import { GenesysCloudMediaSession, IPendingSession } from 'genesys-cloud-streaming-client';
+import { RequestApiOptions, IPendingSession, SessionEvents } from 'genesys-cloud-streaming-client';
+import { RootState, AppDispatch } from "../store.ts";
+import {
+  addParticipantUpdateToVideoConversation,
+  addVideoConversationToActive, removeVideoConversationFromActive,
+  setActiveParticipants, setUsersTalking,
+  updateConversationMediaStreams
+} from "../features/videoConversationsSlice.ts";
+import { useRef } from "react";
 
 interface IAuthData {
   token: string;
@@ -25,10 +33,15 @@ interface IAuthData {
   };
 }
 
+type IMediaTypes = 'audio' | 'video' | 'both'
+
 export default function useSdk() {
   let webrtcSdk: GenesysCloudWebrtcSdk;
-  const dispatch = useDispatch();
-  const sdk = useSelector((state) => state.sdk.sdk);
+  const dispatch = useDispatch<AppDispatch>();
+
+  const memberStatusUpdateRef = useRef<{ memberStatusMessage: MemberStatusMessage }>();
+
+  const sdk: GenesysCloudWebrtcSdk = useSelector((state: RootState) => state.sdk.sdk);
 
   async function initWebrtcSDK(authData: IAuthData) {
     const options: ISdkConfig = {
@@ -79,23 +92,130 @@ export default function useSdk() {
   function handlePendingSession(pendingSession: IPendingSession): void {
     dispatch(updatePendingSessions(pendingSession));
   }
+
   // If a pendingSession was cancelled or handled, we can remove it from our state.
-  function handleCancelPendingSession(pendingSession: IPendingSession): void {
+  function handleCancelPendingSession(pendingSession: ISessionIdAndConversationId): void {
     dispatch(removePendingSession(pendingSession));
   }
 
-  function handledPendingSession(pendingSession: IPendingSession): void {
+  function handledPendingSession(pendingSession: ISessionIdAndConversationId): void {
     dispatch(removePendingSession(pendingSession));
     dispatch(storeHandledPendingSession(pendingSession))
   }
 
-  function handleSessionStarted() {}
+  async function handleSessionStarted(session: IExtendedMediaSession) {
+    console.log("ASDNEGIURHFEJDFHGIUEJOAIDF")
+    if (session.sessionType === 'collaborateVideo') {
 
-  function handleSessionEnded(session: GenesysCloudMediaSession) {}
+      const sessionEventsToLog: Array<keyof SessionEvents> = ['participantsUpdate', 'activeVideoParticipantsUpdate', 'speakersUpdate'];
+      sessionEventsToLog.forEach((eventName) => {
+        session.on(eventName, (e) => console.info(eventName, e));
+      })
 
-  function handleDisconnected() {}
+      const localMediaStream = await webrtcSdk.media.startMedia({ video: true, audio: true });
 
-  function handleConnected() {}
+      webrtcSdk.acceptSession({
+        conversationId: session.conversationId,
+        audioElement: document.createElement('audio'),
+        videoElement: document.createElement('video'),
+        mediaStream: localMediaStream
+      });
+
+      dispatch(addVideoConversationToActive({
+        session: session,
+        conversationId: session.conversationId,
+      }));
+
+      if (session?._outboundStream) {
+        dispatch(updateConversationMediaStreams({
+          conversationId: session.conversationId,
+          outboundStream: session._outboundStream,
+        }));
+      }
+
+      setupSessionListeners(session);
+    }
+  }
+
+  const updateMemberStatus = (memberStatusMessage: MemberStatusMessage, convId: string) => {
+    const lastUpdateParams = memberStatusUpdateRef.current?.memberStatusMessage?.params || {};
+    const mergedUpdateParams = { ...lastUpdateParams, ...memberStatusMessage.params }
+    const mergedUpdate = { ...memberStatusMessage, params: mergedUpdateParams }
+
+    if (memberStatusMessage?.params?.incomingStreams) {
+      const userIds = memberStatusMessage.params.incomingStreams.map(stream => {
+        const appId = stream.appId || stream.appid;
+        return appId?.sourceUserId
+      });
+      dispatch(setActiveParticipants({
+        conversationId: convId,
+        activeParticipants: userIds
+      }));
+    }
+
+    if (memberStatusMessage?.params?.speakers) {
+      const usersTalking = memberStatusMessage.params.speakers.reduce((acc, current) => {
+        return { ...acc, [current.appId.sourceUserId]: current.activity === 'speaking' }
+      }, {});
+      dispatch(setUsersTalking({
+        conversationId: convId,
+        usersTalking
+      }));
+    }
+
+    memberStatusUpdateRef.current = {
+      memberStatusMessage: mergedUpdate
+    };
+  }
+
+  const setupSessionListeners = (session: IExtendedMediaSession) => {
+    // Save the incoming media stream to allow switching between conversations
+    session.on('incomingMedia', () => {
+      if (session.pc.getReceivers) {
+        const receivers = session.pc.getReceivers();
+        const inboundTracks = receivers
+          .map(receiver => receiver.track)
+          .filter(track => track);
+        if (inboundTracks.length > 0) {
+          const inboundStream = new MediaStream(inboundTracks);
+          dispatch(updateConversationMediaStreams({
+            conversationId: session.conversationId,
+            inboundStream: inboundStream,
+          }));
+        }
+      }
+    });
+
+    // Used for mute/unmute, screen share
+    session.on('participantsUpdate', partsUpdate => {
+      dispatch(addParticipantUpdateToVideoConversation(partsUpdate));
+    });
+
+    // Remove conversation from store
+    session.on('terminated', reason => {
+      dispatch(removeVideoConversationFromActive({ conversationId: session.conversationId, reason: reason }));
+    });
+
+    // set green border when talking
+    session.on('memberStatusUpdate', (memberStatusMessage: MemberStatusMessage) => updateMemberStatus(memberStatusMessage, session.conversationId));
+  }
+
+  function startVideoConference(roomJid: string): Promise<{ conversationId: string; }> {
+    return sdk.startVideoConference(roomJid);
+  }
+
+  function startVideoMeeting(roomJid: string): Promise<{ conversationId: string; }> {
+    return sdk.startVideoMeeting(roomJid);
+  }
+
+  function handleSessionEnded(_session: IExtendedMediaSession) {
+  }
+
+  function handleDisconnected() {
+  }
+
+  function handleConnected() {
+  }
 
   function handleConversationUpdate(update: ISdkConversationUpdateEvent): void {
     dispatch(updateConversations(update));
@@ -104,18 +224,25 @@ export default function useSdk() {
   function endSession(conversationId: string): void {
     sdk.endSession({ conversationId });
   }
+
   async function toggleAudioMute(mute: boolean, conversationId: string): Promise<void> {
     await sdk.setAudioMute({ mute, conversationId });
   }
+
+  async function toggleVideoMute(mute: boolean, conversationId: string): Promise<void> {
+    await sdk.setVideoMute({ mute, conversationId });
+  }
+
   async function toggleHoldState(held: boolean, conversationId: string): Promise<void> {
     await sdk.setConversationHeld({ held, conversationId });
   }
 
-  function handleMediaStateChange(state: void): void {
-    dispatch(updateMediaState(state));
+  function handleMediaStateChange(mediaState: SdkMediaStateWithType): void {
+    dispatch(updateMediaState(mediaState));
   }
-  function handleGumRequest(state: void): void {
-    dispatch(updateGumRequests(state));
+
+  function handleGumRequest(_gumRequest: ISdkGumRequest): void {
+    dispatch(updateGumRequests());
   }
 
   function updateDefaultDevices(options: any): void {
@@ -124,13 +251,16 @@ export default function useSdk() {
       updateActiveSessions: true,
     });
   }
+
   function enumerateDevices(): void {
     sdk.media.enumerateDevices(true);
   }
-  function requestDevicePermissions(type: string): void {
+
+  function requestDevicePermissions(type: IMediaTypes): void {
     sdk.media.requestMediaPermissions(type);
   }
-  function updateAudioVolume(volume: string): void {
+
+  function updateAudioVolume(volume: number): void {
     sdk.updateAudioVolume(volume);
   }
 
@@ -142,7 +272,7 @@ export default function useSdk() {
   async function updateOnQueueStatus(onQueue: boolean): Promise<void> {
     const systemPresences = await sdk._http.requestApi(`systempresences`, {
       method: 'get',
-      host: sdk._config.environment,
+      host: sdk._config.environment || 'inindca.com',
       authToken: sdk._config.accessToken
     });
 
@@ -152,9 +282,9 @@ export default function useSdk() {
     } else {
       presenceDefinition = systemPresences.data.find((p: { name: string; }) => p.name === 'AVAILABLE')
     }
-    const requestOptions = {
+    const requestOptions: RequestApiOptions = {
       method: 'patch',
-      host: sdk._config.environment,
+      host: sdk._config.environment || 'inindca.com',
       authToken: sdk._config.accessToken,
       data: JSON.stringify({ presenceDefinition })
     };
@@ -163,8 +293,12 @@ export default function useSdk() {
   }
 
   function disconnectPersistentConnection(): void {
-    const sessions = sdk.sessionManager.getAllActiveSessions().filter((session: GenesysCloudMediaSession) => session.sessionType === SessionTypes.softphone);
-    sessions.forEach((session: GenesysCloudMediaSession) => sdk.forceTerminateSession(session.id));
+    const sessions = sdk.sessionManager.getAllActiveSessions().filter((session: IExtendedMediaSession) => session.sessionType === SessionTypes.softphone);
+    sessions.forEach((session: IExtendedMediaSession) => sdk.forceTerminateSession(session.id));
+  }
+
+  function getSession(conversationId: string): VideoMediaSession {
+    return sdk.sessionManager.getSession({conversationId}) as VideoMediaSession;
   }
 
   return {
@@ -172,6 +306,7 @@ export default function useSdk() {
     startSoftphoneSession,
     endSession,
     toggleAudioMute,
+    toggleVideoMute,
     toggleHoldState,
     updateDefaultDevices,
     enumerateDevices,
@@ -179,6 +314,9 @@ export default function useSdk() {
     updateAudioVolume,
     destroySdk,
     updateOnQueueStatus,
-    disconnectPersistentConnection
+    disconnectPersistentConnection,
+    startVideoConference,
+    startVideoMeeting,
+    getSession
   };
 }