+++ react-demo-app/src/store.ts
@@ -1,6 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import authReducer from './features/authSlice';
 import conversationsReducer from './features/conversationsSlice';
+import videoConversationReducer from './features/videoConversationsSlice';
 import devicesReducer from './features/devicesSlice';
 import sdkReducer from './features/sdkSlice';
 
@@ -9,8 +10,12 @@ export const store= configureStore({
   reducer: {
     auth: authReducer,
     conversations: conversationsReducer,
+    videoConversations: videoConversationReducer,
     devices: devicesReducer,
     sdk: sdkReducer
   },
   middleware: getDefaultMiddleware => getDefaultMiddleware({serializableCheck: false})
 })
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;